Revert "move c++ lstm to also use size_t, since Rust uses usize"
This reverts commit 3b174eadf936169ce9eced5879cf3055ae3d354f.

+++ enzyme/benchmarks/ReverseMode/lstm/lstm.h
@@ -17,9 +17,9 @@ extern "C" {
 //      state (2 * l * b)
 //      sequence (c * b)
 void lstm_objective(
-    size_t l,
-    size_t c,
-    size_t b,
+    int l,
+    int c,
+    int b,
     double const* main_params,
     double const* extra_params,
     double* state,
